Update tasks 2.3 and 2.4

+++ src/module_02/homeWork/task23.java
@@ -1,7 +1,7 @@
 package module_02.homeWork;
 
 /**
- * Created by voksus on 30.01.2017
+ * Created by voksus on 05.02.2017
  */
 public class task23 {
 
@@ -18,7 +18,6 @@ public class task23 {
     }
 
     // method checking owner's balance and trying to withdraw money if possible
-    // disadvantage of this method is that it does not account for decimal values
     private static String withdraw(String owner, double withdrawal) {
         StringBuilder result = new StringBuilder();
         int ownerIndex = -1;
@@ -35,7 +34,7 @@ public class task23 {
         if (ownerIndex >= 0) {
             if (balances[ownerIndex] > 0 && withdrawal > 0 && withdrawal * 1.05 < balances[ownerIndex]) {
                 result.append(ownerNames[ownerIndex]).append(" ");
-                result.append(Double.toString((int) withdrawal)).append(" ");
+                result.append(Double.toString(withdrawal)).append(" ");
                 result.append(Double.toString(balances[ownerIndex] - (withdrawal * 1.05)));
             } else {
                 result.append(owner).append(" NO");
